Add compile_directory.json (#99)
Adds a new JSON representation of the compile directory, which basically
has a list of output artifacts by name. Here's what an example json
would look like:

```
{
  "[0/0]": {
    "artifacts": [
      {
        "name": "dynamo_output_graph_0.txt",
        "number": 0,
        "suffix": "",
        "url": "-_0_0_0/dynamo_output_graph_0.txt"
      },
      {
        "name": "aot_forward_graph_1.txt",
        "number": 1,
        "suffix": "",
        "url": "-_0_0_0/aot_forward_graph_1.txt"
      },
      {
        "name": "fx_graph_cache_hash_2.json",
        "number": 2,
        "suffix": "",
        "url": "-_0_0_0/fx_graph_cache_hash_2.json"
      },
      {
        "name": "dynamo_guards_3.html",
        "number": 3,
        "suffix": "",
        "url": "-_0_0_0/dynamo_guards_3.html"
      },
      {
        "name": "dynamo_cpp_guards_str_4.txt",
        "number": 4,
        "suffix": "",
        "url": "-_0_0_0/dynamo_cpp_guards_str_4.txt"
      },
      {
        "name": "compilation_metrics_5.html",
        "number": 5,
        "suffix": "",
        "url": "-_0_0_0/compilation_metrics_5.html"
      }
    ]
  }
}
```

Today, name and url are actually the same for basically all results. But
that might change in the future and it's more generic to just dump the
entire JSON. This allows external programs to understand the directory
hierarchy and contents of the tlparse at a glance, without having to
parse index.html