Support runtime compile ids which don't have an attempt (#101)
Runtime autotuning compile ids don't have an attempt associated with
them, which results in duplicate rows in the compile directory. This
fixes this issue.

This does *not* implement support for runtime_compilation_metrics, which
I'll leave as a TODO for when we want to build support for that in
tlparse.

Added a new test log, along with the result, where we now see no
duplicate [0/0]:
<img width="1052" alt="image"
src="https://github.com/user-attachments/assets/49107644-921e-4969-882e-3e86beb9e08d"
/>